Add optional precedence parameter to addBinaryOp.
Using jsep.addBinaryOp without precedence causes it to fail to
parse some strings, so this package needs to make sure to always
pass a precedence value. For backwards compatibility, if
precedence is omitted it's set to either the original precedence
of that operator (when overriding a built-in operator) or 1 (when
defining a new operator).