Combined sync/async evaluators!
After seeing how so nearly identical the sync/async evaluators were, I converted to a class in order to maintain an `isAsync` property while parsing. In doing so, I also moved the Context argument into the class instance instead of passing into all evaluation methods.

Now there's a new `evalSyncAsync` method which takes a callback method so that it can either call it back immediately with the value (sync), or in a promise chain (async). (Saves more than 20% of loc)